+++ config/routes.rb
@@ -1,5 +1,3 @@
-# frozen_string_literal: true
-
 ActiveadminMcp::Engine.routes.draw do
   post "/", to: "mcp#call"
 end