+++ spec/e2e/fixture_app/app/models/author.rb
@@ -1,0 +1,6 @@
+class Author < ApplicationRecord
+  # See the note in post.rb: Ransack 4 requires an explicit allowlist.
+  def self.ransackable_attributes(_auth_object = nil)
+    column_names
+  end
+end