+++ spec/spec_helper.rb
@@ -1,5 +1,3 @@
-# frozen_string_literal: true
-
 require "rails"
 require "active_record"
 require "action_controller"